Move ColorPicker implementation and remove dangling Gradient extension
