--- Sources/GtkBackend/GtkBackend+Gradient.swift
@@ -1,8 +1,0 @@
-import CGtk
-import Gtk
-@_spi(Backends) import SwiftCrossUI
-import Foundation
-
-extension GtkBackend {
-
-}
